Rename rocketmq-client to rocketmq
