+++ python/tests/test_foo.py
@@ -1,4 +1,4 @@
-from rocketmq_client import foo
+from rocketmq import foo
 
 def test_passing():
     assert (1, 2, 3) == (1, 2, 3)
